Move helper functions to own files
